feat: print slides --with-clicks option
